Housekeeping changes
Added a txt extension to the license file. Moved all source files into the src folder. Edited the Makefile so it will now access source files from the src folder. Also added a debug target to the Makefile which can be used to generate debug builds separate of release builds.